Props moved out
