--- src/webparts/workMinders/components/IWorkMindersProps.ts
@@ -1,12 +1,0 @@
-import { WebPartContext } from "@microsoft/sp-webpart-base";
-import { TWorkMinder } from "../types/ItemTypes";
-
-export interface IWorkMindersProps {
-  isDarkTheme: boolean;
-  hasTeamsContext: boolean;
-  webpartContext: WebPartContext;
-  workMinders: TWorkMinder[];
-  height: number;
-  smallUi: boolean;
-  oneDriveDoesNotExist: boolean;
-}
